Move artifactReducer into it's own file and add test cases
